requirements.txt
google-ads==21.3.0